+++ lib/v2/screens/dashboard/components/tokens_cards/tokens_cards.dart
@@ -9,7 +9,7 @@ import 'package:seeds/v2/domain-shared/page_state.dart';
 import 'package:seeds/v2/screens/dashboard/components/tokens_cards/interactor/viewmodels/bloc.dart';
 import 'package:seeds/v2/screens/dashboard/interactor/viewmodels/bloc.dart';
 
-import '../currency_info_card_widget.dart';
+import 'currency_info_card_widget.dart';
 
 class TokenCards extends StatefulWidget {
   const TokenCards({Key? key}) : super(key: key);
