+++ lib/v2/screens/dashboard/components/transactions_list/interactor/mappers/transactions_state_mapper.dart
@@ -1,6 +1,6 @@
-import 'package:seeds/v2/domain-shared/page_state.dart';
-import 'package:seeds/v2/screens/dashboard/interactor/viewmodels/transactions_list_state.dart';
 import 'package:async/async.dart';
+import 'package:seeds/v2/domain-shared/page_state.dart';
+import '../viewmodels/transactions_list_state.dart';
 
 class TransactionsListStateMapper {
   TransactionsListState mapResultToState(TransactionsListState currentState, Result transactionsResult) {
